Created new enum for scanType
